redoing the seperation by wrapping widgets editor components
